+++ lib/src/controllers/types/mp_setting_type.dart
@@ -26,7 +26,7 @@ enum MPSettingID {
   TH2Edit_LineThickness,
   TH2Edit_NewLineCreationMethod,
   TH2Edit_PointRadius,
-  TH2Edit_RotationSnapAngle,
+  TH2Edit_SnapAngle,
   TH2Edit_SelectionTolerance,
   TH2Edit_ShowDirectionTicksOnNonSelectedLines;
 
@@ -53,7 +53,7 @@ enum MPSettingID {
     MPSettingID.TH2Edit_LineThickness: MPSettingType.double,
     MPSettingID.TH2Edit_NewLineCreationMethod: MPSettingType.enumeration,
     MPSettingID.TH2Edit_PointRadius: MPSettingType.double,
-    MPSettingID.TH2Edit_RotationSnapAngle: MPSettingType.double,
+    MPSettingID.TH2Edit_SnapAngle: MPSettingType.double,
     MPSettingID.TH2Edit_SelectionTolerance: MPSettingType.double,
     MPSettingID.TH2Edit_ShowDirectionTicksOnNonSelectedLines:
         MPSettingType.bool,